Stupid jumping platform puzzle wasting my time
Spent way to much freaking time on this. What a waste...

+++ Bear Prototypes/Assets/Scripts/PlayerScripts/MoveChar.cs
@@ -20,7 +20,7 @@ private float speed;
 private float gravity;
 public float jumpHeight; 
 private float maxJump;
-public float jumpCount;
+public int jumpCount;
 public float maxFallSpeed = -30;
 public float curFallSpeed;
 Quaternion myRotate;
